test(finite-geometry): own catalog and dispatch coverage in the dispatch lane
